Point Cloud VFX Binding
Main update to automate the process of creating new point clouds and vfx.
Mainly consists of 2 updates
1- VFX Graph: Exposed VFX properties. This allows 1) To be able to play around with values from the inspector. 2) To allow multiple point clouds to share the same VFX Asset with different values. 3) Animating VFX Parameters in timeline.
2- Add PointCloud & PointCloudBinder scripts: This simplifies the process of adding new point clouds to the scene, to one simple click, by simply selecting a different point cloud in the inspector, by binding the properties in the VFX Graph to the baked point cloud asset,. The binded properties are 1) Position & Color Maps 2) Spawn Rate 3) Capacity 4) Game Object Name.

There is a 3rd useful step not included in the changes. It updates PointClouds container type by default to Texture. For this:
Update in Packages > PCX Editor > PlyImporter.cs line 23 to [SerializeField] ContainerType _containerType = ContainerType.Texture;